+++ package.json
@@ -16,11 +16,11 @@
     "test": "bin/test --no-oracle",
     "testsqlite3": "NODE_ENV=test ava test/*.js",
     "testpg": "NODE_ENV=test DB=PG ava test/*.js",
-    "testpg-paging": "NODE_ENV=test DB=PG PAGINATE=keyset ava test/pagination/keyset-paging.js && NODE_ENV=test DB=PG PAGINATE=offset ava test/pagination/offset-paging.js",
+    "testpg-paging": "NODE_ENV=test DB=PG PAGINATE=offset ava test/pagination/offset-paging.js test/pagination/common && NODE_ENV=test DB=PG PAGINATE=keyset ava test/pagination/keyset-paging.js test/pagination/common",
     "testoracle": "NODE_ENV=test DB=ORACLE ava test/*.js",
-    "testoracle-paging": "NODE_ENV=test DB=ORACLE PAGINATE=offset ava test/pagination/offset-paging.js && NODE_ENV=test DB=ORACLE PAGINATE=keyset ava test/pagination/keyset-paging.js",
+    "testoracle-paging": "NODE_ENV=test DB=ORACLE PAGINATE=offset ava test/pagination/offset-paging.js test/pagination/common && NODE_ENV=test DB=ORACLE PAGINATE=keyset ava test/pagination/keyset-paging.js test/pagination/common",
     "testmysql": "NODE_ENV=test DB=MYSQL ava test/*.js",
-    "testmysql-paging": "NODE_ENV=test DB=MYSQL PAGINATE=offset ava test/pagination/offset-paging.js && NODE_ENV=test DB=MYSQL PAGINATE=keyset ava test/pagination/keyset-paging.js",
+    "testmysql-paging": "NODE_ENV=test DB=MYSQL PAGINATE=offset ava test/pagination/offset-paging.js test/pagination/common && NODE_ENV=test DB=MYSQL PAGINATE=keyset ava test/pagination/keyset-paging.js test/pagination/common",
     "testtsd": "npm run build && tsd",
     "coverage": "nyc --reporter=html npm run test",
     "view-coverage": "open coverage/index.html",
