+++ test/pagination/keyset-paging.js
@@ -735,93 +735,6 @@ test('should handle "where" condition on main table of many-to-many relation', a
   t.deepEqual(expect, data)
 })
 
-test('should handle order columns on the main table', async t => {
-  const source  = `{
-    user(id: 2) {
-      fullName
-      following(first: 2, sortOnMain: true, after: "${objToCursor({
-        created_at: '2015-10-19T05:48:04.537Z',
-        id: 3
-      })}") {
-        edges {
-          node {
-            id
-            fullName
-          }
-        }
-      }
-    }
-  }`
-  const { data, errors } = await graphql({schema, source})
-  errCheck(t, errors)
-  const expect = {
-    user: {
-      fullName: 'Hudson Hyatt',
-      following: {
-        edges: [
-          {
-            node: {
-              id: toGlobalId('User', 1),
-              fullName: 'Alivia Waelchi'
-            }
-          },
-          {
-            node: {
-              id: toGlobalId('User', 2),
-              fullName: 'Hudson Hyatt'
-            }
-          }
-        ]
-      }
-    }
-  }
-  t.deepEqual(expect, data)
-})
-
-test('should handle order columns on the junction table', async t => {
-  const cursor = objToCursor({
-    created_at: '2016-01-01T16:28:00.051Z',
-    followee_id: 1
-  })
-  const source  = `{
-    user(id: 2) {
-      fullName
-      following(first: 2, sortOnMain: false, after: "${cursor}") {
-        edges {
-          node {
-            id
-            fullName
-          }
-        }
-      }
-    }
-  }`
-  const { data, errors } = await graphql({schema, source})
-  errCheck(t, errors)
-  const expect = {
-    user: {
-      fullName: 'Hudson Hyatt',
-      following: {
-        edges: [
-          {
-            node: {
-              id: toGlobalId('User', 3),
-              fullName: 'Coleman Abernathy'
-            }
-          },
-          {
-            node: {
-              id: toGlobalId('User', 2),
-              fullName: 'Hudson Hyatt'
-            }
-          }
-        ]
-      }
-    }
-  }
-  t.deepEqual(expect, data)
-})
-
 test('should handle an interface type', async t => {
   const source  = `{
     user(id: 1) {
