+++ test/pagination/offset-paging.js
@@ -661,86 +661,6 @@ test('should handle "where" condition on main table of many-to-many relation', a
   t.deepEqual(expect, data)
 })
 
-test('should handle order columns on the main table', async t => {
-  const source = `{
-    user(id: 2) {
-      fullName
-      following(first: 2, sortOnMain: true, after: "${offsetToCursor(0)}") {
-        edges {
-          node {
-            id
-            fullName
-          }
-        }
-      }
-    }
-  }`
-  const { data, errors } = await graphql({schema, source})
-  errCheck(t, errors)
-  const expect = {
-    user: {
-      fullName: 'Hudson Hyatt',
-      following: {
-        edges: [
-          {
-            node: {
-              id: toGlobalId('User', 1),
-              fullName: 'Alivia Waelchi'
-            }
-          },
-          {
-            node: {
-              id: toGlobalId('User', 2),
-              fullName: 'Hudson Hyatt'
-            }
-          }
-        ]
-      }
-    }
-  }
-  t.deepEqual(expect, data)
-})
-
-test('should handle order columns on the junction table', async t => {
-  const source = `{
-    user(id: 2) {
-      fullName
-      following(first: 2, sortOnMain: false, after: "${offsetToCursor(0)}") {
-        edges {
-          node {
-            id
-            fullName
-          }
-        }
-      }
-    }
-  }`
-  const { data, errors } = await graphql({schema, source})
-  errCheck(t, errors)
-  const expect = {
-    user: {
-      fullName: 'Hudson Hyatt',
-      following: {
-        edges: [
-          {
-            node: {
-              id: toGlobalId('User', 2),
-              fullName: 'Hudson Hyatt'
-            }
-          },
-          {
-            node: {
-              id: toGlobalId('User', 3),
-              fullName: 'Coleman Abernathy'
-            }
-          }
-        ]
-      }
-    }
-  }
-  t.deepEqual(expect, data)
-})
-
 test('should handle an interface type', async t => {
   const source = `{
     user(id: 1) {
